+++ examples/sdk-app-example/src/javaExample/java/com/bugsnag/android/example/ExampleActivity.java
@@ -17,6 +17,7 @@ import android.net.Uri;
 import android.os.Bundle;
 import android.support.v7.app.ActionBar;
 import android.support.v7.app.AppCompatActivity;
+import android.util.Log;
 import android.view.View;
 import android.widget.Toast;
 
@@ -52,6 +53,17 @@ public class ExampleActivity extends AppCompatActivity {
             @Override
             public void onClick(View view) {
                 doCrash();
+            }
+        });
+
+        findViewById(R.id.btn_anr).setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                try {
+                    Thread.sleep(10000);
+                } catch (InterruptedException ignored) {
+                    Log.d("Bugsnag", "Interrupted");
+                }
             }
         });
     }