+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ### Enhancements
 
+* Add ANR detection to bugsnag-android
+[#442](https://github.com/bugsnag/bugsnag-android/pull/442)
+
 * Add unhandled_events field to native payload
 [#445](https://github.com/bugsnag/bugsnag-android/pull/445)